+++ anvil_consortium_manager/forms.py
@@ -9,7 +9,6 @@ from django.conf import settings
 from django.core.exceptions import ValidationError
 
 from . import models
-from .auditor.models import IgnoredManagedGroupMembership
 
 
 class Bootstrap5MediaFormMixin:
@@ -334,27 +333,3 @@ class WorkspaceGroupSharingForm(Bootstrap5MediaFormMixin, forms.ModelForm):
                 attrs={"data-theme": "bootstrap-5"},
             ),
         }
-
-
-class IgnoredManagedGroupMembershipForm(Bootstrap5MediaFormMixin, forms.ModelForm):
-    """Form for the IgnoredManagedGroupMembership model."""
-
-    group = forms.ModelChoiceField(
-        queryset=models.ManagedGroup.objects.filter(is_managed_by_app=True),
-        help_text="Only groups managed by this app can be selected.",
-        widget=autocomplete.ModelSelect2(
-            url="anvil_consortium_manager:managed_groups:autocomplete",
-            attrs={
-                "data-theme": "bootstrap-5",
-            },
-            forward=(forward.Const(True, "only_managed_by_app"),),
-        ),
-    )
-
-    class Meta:
-        model = IgnoredManagedGroupMembership
-        fields = (
-            "group",
-            "ignored_email",
-            "note",
-        )
